pyproject.toml
[project]
name = "movies"
version = "0.1.0"
description = ""
authors = [
    {name = "evavallistu",email = "evavallistu@gmail.com"}
]
#readme = "README.md"
requires-python = ">=3.11, <4"
dependencies = [
    "anyio (==3.7.1)",
    "bcrypt (==4.0.1)",
    "cffi (==1.15.1)",
    "click (==8.1.7)",
    "colorama (==0.4.6)",
    "cryptography (==41.0.3)",
    "ecdsa (==0.18.0)",
    "fastapi (==0.101.1)",
    "greenlet (==2.0.2)",
    "h11 (==0.14.0)",
    "idna (==3.4)",
    "passlib (==1.7.4)",
    "psycopg2 (==2.9.10)",
    "pyasn1 (==0.5.0)",
    "pycparser (==2.21)",
    "python-jose (==3.3.0)",
    "python-multipart (==0.0.6)",
    "rsa (==4.9)",
    "six (==1.16.0)",
    "sniffio (==1.3.0)",
    "sqlalchemy (==2.0.20)",
    "starlette (==0.27.0)",
    "typing-extensions (>=4.12.2)",
    "uvicorn (==0.23.2)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)"
]

[tool.poetry]
packages = [{include = "movies"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
